07_databricks_app/requirements.txt
streamlit
psycopg[binary,pool]
databricks-sdk
pandas
plotly
